Move Header component to use it globally
