+++ pages/_app.js
@@ -2,6 +2,10 @@ import "../styles/globals.css";
 import { ClerkProvider, SignedIn, SignedOut } from "@clerk/clerk-react";
 import { useRouter } from "next/router";
 import { useEffect } from "react";
+import styles from "../styles/Home.module.css";
+import Image from "next/image";
+import Link from "next/link";
+import { UserButton } from "@clerk/clerk-react";
 
 const clerkClientHost = process.env.NEXT_PUBLIC_CLERK_HOST;
 
@@ -14,6 +18,32 @@ const clerkClientHost = process.env.NEXT_PUBLIC_CLERK_HOST;
  *  "/foo/[...bar]"  for pages/foo/[...bar].js
  */
 const publicPages = ["/", "/sign-in/[[...index]]", "/sign-up/[[...index]]"];
+
+// Header component using <SignedIn> & <SignedOut>.
+//
+// The SignedIn and SignedOut components are used to control rendering depending
+// on whether or not a visitor is signed in.
+//
+// https://docs.clerk.dev/frontend/react/signedin-and-signedout
+export const Header = () => (
+  <header className={styles.header}>
+    <div className={styles.left}>
+      <Link href="/">
+        <a className={styles.logo}>
+          <Image src="/clerk-logo.svg" width="108" height="32" alt="Clerk" />
+        </a>
+      </Link>
+    </div>
+    <div className={styles.right}>
+      <SignedOut>
+        <Link href="/sign-in">Sign in</Link>
+      </SignedOut>
+      <SignedIn>
+        <UserButton userProfileURL="/user" afterSignOutAll="/" />
+      </SignedIn>
+    </div>
+  </header>
+);
 
 function MyApp({ Component, pageProps }) {
   const router = useRouter();
@@ -28,10 +58,14 @@ function MyApp({ Component, pageProps }) {
       navigate={(to) => router.push(to)}
     >
       {publicPages.includes(router.pathname) ? (
-        <Component {...pageProps} />
+        <>
+          <Header />
+          <Component {...pageProps} />
+        </>
       ) : (
         <>
           <SignedIn>
+            <Header />
             <Component {...pageProps} />
           </SignedIn>
           <SignedOut>
