+++ pages/user/[[...index]].js
@@ -1,10 +1,4 @@
 import { UserProfile } from "@clerk/clerk-react";
-import { Header } from "..";
 export default function UserProfilePage() {
-  return (
-    <>
-      <Header />
-      <UserProfile path="/user" routing="path" />
-    </>
-  );
+  return <UserProfile path="/user" routing="path" />;
 }
